Add stack implementation and virtual machine code
This commit adds the implementation of a stack data structure in the stack.js file. It also includes the initial implementation of a virtual machine in the vm.js file. The virtual machine can execute instructions such as PUSH, POP, ADD, SUB, MUL, DIV, SWAP, EMPTY, and PRINT. Additionally, it can read instructions from a file and execute all .val files in a given directory.